requirements.txt
pydantic>=2.0
google-genai>=1.0
python-dotenv>=1.0
ollama>=0.3.0
pypdf>=4.0
psycopg2-binary>=2.9
